+++ integration/spring-hateoas/webmvc/src/test/java/com/blazebit/persistence/spring/hateoas/webmvc/config/BlazePersistenceConfiguration.java
@@ -1,0 +1,76 @@
+/*
+ * Copyright 2014 - 2020 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.spring.hateoas.webmvc.config;
+
+import com.blazebit.persistence.Criteria;
+import com.blazebit.persistence.CriteriaBuilderFactory;
+import com.blazebit.persistence.spi.CriteriaBuilderConfiguration;
+import com.blazebit.persistence.view.EntityViewManager;
+import com.blazebit.persistence.view.spi.EntityViewConfiguration;
+import com.blazebit.persistence.view.spi.type.TypeConverter;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.context.annotation.Scope;
+
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.PersistenceUnit;
+import java.lang.reflect.Type;
+import java.util.UUID;
+
+/**
+ * @author Moritz Becker
+ * @since 1.2.0
+ */
+@Configuration
+public class BlazePersistenceConfiguration {
+
+    @PersistenceUnit
+    private EntityManagerFactory entityManagerFactory;
+
+    @Bean
+    @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
+    @Lazy(false)
+    public CriteriaBuilderFactory createCriteriaBuilderFactory() {
+        CriteriaBuilderConfiguration config = Criteria.getDefault();
+        return config.createCriteriaBuilderFactory(entityManagerFactory);
+    }
+
+    @Bean
+    @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
+    @Lazy(false)
+    public EntityViewManager createEntityViewManager(CriteriaBuilderFactory cbf, EntityViewConfiguration entityViewConfiguration) {
+        entityViewConfiguration.registerTypeConverter(String.class, UUID.class, new TypeConverter<String, UUID>() {
+            @Override
+            public Class<?> getUnderlyingType(Class<?> owningClass, Type declaredType) {
+                return String.class;
+            }
+
+            @Override
+            public UUID convertToViewType(String object) {
+                return object == null ? null : UUID.fromString(object);
+            }
+
+            @Override
+            public String convertToUnderlyingType(UUID object) {
+                return object == null ? null : object.toString();
+            }
+        });
+        return entityViewConfiguration.createEntityViewManager(cbf);
+    }
+}